extract helper to create mock redux store
